On review; fix warning:
- Reference to property 'stompClient' in closure requires explicit use of 'self' to make capture semantics explicit; this is an error in Swift 6
