+++ ArtemisKit/Sources/Messages/Views/MessageDetailView/MessageDetailView.swift
@@ -30,24 +30,6 @@ struct MessageDetailView: View {
         self._message = message
     }
 
-    init(viewModel: ConversationViewModel, messageId: Int64) {
-        self.viewModel = viewModel
-        self.messageId = messageId
-        self._message = Binding(get: { .loading }, set: { _ in return })
-        self.internalMessage = nil
-
-        self._message = Binding(get: { [self] in
-            if let internalMessage = self.internalMessage {
-                return .done(response: internalMessage)
-            }
-            return .loading
-        }, set: { [self] in
-            if let message = $0.value as? Message {
-                self.internalMessage = message
-            }
-        })
-    }
-
     var body: some View {
         DataStateView(data: $message) {
             await reloadMessage()
@@ -130,11 +112,10 @@ private extension MessageDetailView {
                     $0.creationDate ?? .tomorrow < $1.creationDate ?? .yesterday
                 }
                 ForEach(Array(sortedArray.enumerated()), id: \.1) { index, answerMessage in
-                    let showHeader = index == 0 || !answerMessage.isContinuation(of: sortedArray[index - 1])
                     MessageCellWrapper(
                         viewModel: viewModel,
                         answerMessage: answerMessage,
-                        isHeaderVisible: showHeader)
+                        isHeaderVisible: index == 0 || !answerMessage.isContinuation(of: sortedArray[index - 1]))
                 }
                 Spacer()
                     .id("bottom")
@@ -270,112 +251,4 @@ private struct MessageCellWrapper: View {
             message: Binding<DataState<BaseMessage>>.constant(
                 DataState<BaseMessage>.done(response: message)))
     }()
-}
-
-private struct MockMessagesService: MessagesService {
-    let messages: [Message]
-
-    func getConversations(for courseId: Int) async -> Common.DataState<[SharedModels.Conversation]> {
-        .loading
-    }
-
-    func updateIsConversationFavorite(for courseId: Int, and conversationId: Int64, isFavorite: Bool) async -> NetworkResponse {
-        .loading
-    }
-
-    func updateIsConversationMuted(for courseId: Int, and conversationId: Int64, isMuted: Bool) async -> NetworkResponse {
-        .loading
-    }
-
-    func updateIsConversationHidden(for courseId: Int, and conversationId: Int64, isHidden: Bool) async -> NetworkResponse {
-        .loading
-    }
-
-    func getMessages(for courseId: Int, and conversationId: Int64, size: Int) async -> Common.DataState<[SharedModels.Message]> {
-        .done(response: messages)
-    }
-
-    func sendMessage(for courseId: Int, conversation: SharedModels.Conversation, content: String) async -> Common.NetworkResponse {
-        .loading
-    }
-
-    func sendAnswerMessage(for courseId: Int, message: SharedModels.Message, content: String) async -> Common.NetworkResponse {
-        .loading
-    }
-
-    func deleteMessage(for courseId: Int, messageId: Int64) async -> Common.NetworkResponse {
-        .loading
-    }
-
-    func deleteAnswerMessage(for courseId: Int, anserMessageId: Int64) async -> Common.NetworkResponse {
-        .loading
-    }
-
-    func editMessage(for courseId: Int, message: SharedModels.Message) async -> Common.NetworkResponse {
-        .loading
-    }
-
-    func editAnswerMessage(for courseId: Int, answerMessage: SharedModels.AnswerMessage) async -> Common.NetworkResponse {
-        .loading
-    }
-
-    func addReactionToAnswerMessage(for courseId: Int, answerMessage: SharedModels.AnswerMessage, emojiId: String) async -> Common.NetworkResponse {
-        .loading
-    }
-
-    func addReactionToMessage(for courseId: Int, message: SharedModels.Message, emojiId: String) async -> Common.NetworkResponse {
-        .loading
-    }
-
-    func removeReactionFromMessage(for courseId: Int, reaction: SharedModels.Reaction) async -> Common.NetworkResponse {
-        .loading
-    }
-
-    func getChannelsOverview(for courseId: Int) async -> Common.DataState<[SharedModels.Channel]> {
-        .loading
-    }
-
-    func addMembersToChannel(for courseId: Int, channelId: Int64, usernames: [String]) async -> Common.NetworkResponse {
-        .loading
-    }
-
-    func removeMembersFromChannel(for courseId: Int, channelId: Int64, usernames: [String]) async -> Common.NetworkResponse {
-        .loading
-    }
-
-    func addMembersToGroupChat(for courseId: Int, groupChatId: Int64, usernames: [String]) async -> Common.NetworkResponse {
-        .loading
-    }
-
-    func removeMembersFromGroupChat(for courseId: Int, groupChatId: Int64, usernames: [String]) async -> Common.NetworkResponse {
-        .loading
-    }
-
-    func createChannel(for courseId: Int, name: String, description: String?, isPrivate: Bool, isAnnouncement: Bool) async -> Common.DataState<SharedModels.Channel> {
-        .loading
-    }
-
-    func searchForUsers(for courseId: Int, searchText: String) async -> Common.DataState<[SharedModels.ConversationUser]> {
-        .loading
-    }
-
-    func createGroupChat(for courseId: Int, usernames: [String]) async -> Common.DataState<SharedModels.GroupChat> {
-        .loading
-    }
-
-    func createOneToOneChat(for courseId: Int, usernames: [String]) async -> Common.DataState<SharedModels.OneToOneChat> {
-        .loading
-    }
-
-    func getMembersOfConversation(for courseId: Int, conversationId: Int64, page: Int) async -> Common.DataState<[SharedModels.ConversationUser]> {
-        .loading
-    }
-
-    func archiveChannel(for courseId: Int, channelId: Int64) async -> Common.NetworkResponse {
-        .loading
-    }
-
-    func unarchiveChannel(for courseId: Int, channelId: Int64) async -> Common.NetworkResponse {
-        .loading
-    }
 }
